Version4.1 Fix some issues on menu
