+++ Project/Assets/Resources/Saved/5.data.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 786512a3743ea438dad2a752e9be7a87
+timeCreated: 1527069091
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
