+++ Project/Assets/Resources/Saved/6.data.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: daa1425a15c784f33ba24e5ace5e831d
+timeCreated: 1527069091
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
