+++ Project/Assets/Resources/Saved/da1.data.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 096c207ddf31e4ee2839f28ccf0da41e
+timeCreated: 1527069091
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
