--- Project/Assets/Scenes/Test.unity.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 965af43bb595e44a09ccdf0035bf7f32
-timeCreated: 1527003275
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
