--- Project/Assets/Scenes/empty.unity.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: ce9837cbeb3304508a03e53d66e5078d
-timeCreated: 1527003275
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
