--- Project/Assets/Scripts/Move.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 4a659478ee6c945b4a20c35afe0c8ba1
-timeCreated: 1527003275
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
