+++ Project/Library/LastSceneManagerSetup.txt
@@ -1,4 +1,4 @@
 sceneSetups:
-- path: Assets/Scenes/Game.unity
+- path: Assets/Scenes/MainScene.unity
   isLoaded: 1
   isActive: 1
