+++ Project/Mds.userprefs
@@ -1,11 +1,11 @@
 ﻿<Properties>
   <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug" />
-  <MonoDevelop.Ide.Workbench ActiveDocument="Assets/Scripts/Player.cs">
+  <MonoDevelop.Ide.Workbench ActiveDocument="Assets/Scripts/MainMenu/MenuSystem.cs">
     <Files>
-      <File FileName="Assets/Scripts/MainMenu/MenuSystem.cs" Line="190" Column="88" />
-      <File FileName="Assets/Scripts/Player.cs" Line="61" Column="5" />
+      <File FileName="Assets/Scripts/MainMenu/MenuSystem.cs" Line="6" Column="14" />
+      <File FileName="Assets/Scripts/Player.cs" Line="48" Column="46" />
       <File FileName="Assets/Scripts/Terrain/Land.cs" Line="1" Column="1" />
-      <File FileName="Assets/Scripts/GameSystem.cs" Line="33" Column="3" />
+      <File FileName="Assets/Scripts/GameSystem.cs" Line="1" Column="1" />
     </Files>
   </MonoDevelop.Ide.Workbench>
   <MonoDevelop.Ide.DebuggingService.Breakpoints>
